requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djongo==1.3.7
pymongo==3.11.0
python-dotenv==1.0.1
